+++ HISTORY.rst
@@ -3,7 +3,15 @@
 History
 -------
 
-0.1.43
+0.1.43 (2018-04-??)
+______________________
+
+* Fixed wxPython exception import.
+* Use the datasheet link information from KiCad and other EDAs, given by 'datasheet' field.
+* Now automatically merge 'description' and other fields to create the groups.
+
+
+0.1.43 (2018-03-15)
 ______________________
 
 * Fixed RS scrape module.